+++ test/index.spec.js
@@ -115,7 +115,7 @@ test('bissle >> append all passed query parameters', t => {
   const { host, server } = setup();
 
   server.inject('/?page=4&per_page=3&fields=foo', response => {
-    t.equal(response.result._links.first, `/?per_page=3&fields=foo`);
+    t.equal(response.result._links.first.href, `/?per_page=3&fields=foo`);
     t.end();
   });
 });